add conversion webhook (#279)

+++ hack/copy-crds.sh
@@ -4,7 +4,12 @@ source "$(dirname "${BASH_SOURCE}")/init.sh"
 
 for f in $HUB_CRD_FILES
 do
-    cp $f ./manifests/cluster-manager/hub/
+    if [ -f "$PATCHED_DIR/$(basename $f)" ]
+    then
+        cp $PATCHED_DIR/$(basename $f) ./manifests/cluster-manager/hub/
+    else 
+        cp $f ./manifests/cluster-manager/hub/
+    fi
 done
 
 for f in $SPOKE_CRD_FILES
